Use start/end anchors in matchStrings() regex
This adds a new test case demonstrating the issue where "com/**" would incorrectly match "com2/foo" or "2com/foo".

The change also necessitated revising an existing test case, since "t?st" should not match a string with a file extension. The original test case was changed to add an extension to the pattern, and the extension-less pattern was preserved for a new regression test.